usertest: structural refactor — one module per surface under test
Split base/usertest/src/main.rs (3408 lines, omnibus) into a layered
module tree:

  main.rs          — entry + child-mode dispatch + runner driver +
                     ALL TESTS PASSED + shutdown (~60 lines)
  bootstrap.rs     — typed Caps from the creator-endpoint round (the
                     root-filesystem cap is FS-agnostic; the driver
                     behind it is opaque)
  runner.rs        — `Phase { name, run: fn(&Caps) }` + uniform
                     `[usertest] phase=<name> starting/passed` envelope
  reentry.rs       — argv-driven child-mode dispatch into the phase
                     modules that own each role
  ipc_util/{ns,fs,time}.rs — already-extracted raw-IPC helpers
  phases/<surface>.rs — 12 phase modules, one per services-tier or
                     kernel/runtime surface under test:
                       startup, memmgr, threading, timed, shmem,
                       pipes, namespace, fs_ipc, fs_std, procmgr,
                       process_faults, pwrmgr

The composition order in phases::all() preserves the original test
ordering (fs_open_relative_phase last; pwrmgr_shutdown explicitly
post-runner). Behaviour unchanged.

`phases/mod.rs` rustdoc codifies the durable rule: one module per
services-tier surface under test; new service ⇒ new module; files
don't grow unboundedly because each is scoped to one surface.

Verified: cargo xtask build + run on both x86_64 and riscv64;
[usertest] ALL TESTS PASSED on both.

Closes the Part A scope of #21. Part B (svcmgr-owned launch +
config-driven supervision) follows in subsequent commits.
